+++ CHANGELOG.md
@@ -31,6 +31,7 @@ All notable changes to `homebridge-config-ui-x` will be documented in this file.
 - update `@homebridge/node-pty-prebuilt-multiarch` to `0.14.1-beta`
 - added `arm` runners to the ui workflows for testing
 - make 'keep orphans' setting available in ui settings service
+- improve types for widgets
 
 ### Homebridge Dependencies